+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add TOC for single resource PDFs
 - Refactored Playwright/Cucumber tests to use new dropdown action function
 - Refactored Playwright/Cucumber tests to use new text fill function
+- Hide Header and Footer components on /read pages using the Web Reader
 - Migrate from Google Analytics to Adobe Analytics
 
 ## [0.17.1]